Fixed bug where it would short instead of long when it detected a long ema cross.

+++ bot-1.js
@@ -104,7 +104,7 @@ async function hasEMACrossedInMarket(marketId, timeframe) {
       return 'short';
     } else {
       console.log('GO LONG');
-      return 'short';
+      return 'long';
     }
   }
   return;
